refactor: split out data tree
